Require helpers explicitly, move generateTemplate to its own file.
