+++ docs/scopes.markdown
@@ -1,6 +1,6 @@
 [](root)
 ```coffeescript
-require("../test/helper")
+require("../test/support/helper")
 Model = require("Backbone").Model
 EndDash = require("../lib/end-dash").generateTemplate
 ```
