+++ test/backbone_integration.js
@@ -1,7 +1,9 @@
+require('./support/helper');
+
 var expect = require("expect.js")
   , fs = require("fs")
   , Backbone = require("backbone")
-  , generateTemplate = require("./util").generateTemplate
+  , generateTemplate = require("./support/generate_template")
   , Factory = require("test-things").Factory
 
 describe("when integrating with backbone", function() {
