+++ test/conditional_tag.js
@@ -1,9 +1,10 @@
+require('./support/helper');
+
 var path = require("path")
   , expect = require("expect.js")
   , fs = require("fs")
-  , helper = require("./helper")
   , Backbone = require("backbone")
-  , generateTemplate = require("./util").generateTemplate
+  , generateTemplate = require("./support/generate_template")
 
 describe("A conditional tag", function() {
   beforeEach(function() {
