+++ test/model_change.js
@@ -1,11 +1,12 @@
-var generateTemplate = require("./util").generateTemplate
+require('./support/helper');
+
+var generateTemplate = require("./support/generate_template")
   , ViewReaction = require("../lib/reactions/view")
   , Backbone = require("backbone")
   , expect = require("expect.js")
   , _ = require("underscore")
   , fs = require("fs")
   , views = { }
-
 
 describe("When I replace an embedded model", function() {
   beforeEach(function() {
